requirements.txt
llama-cpp-python==0.2.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
